# ONGConnect/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ONGConnect — Match Voluntários e ONGs",
        "description": "API REST para conexão entre ONGs do Alto Vale do Itajaí e voluntários da comunidade. Projeto de Extensão — Unidavi 2026.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/inscricoes/{id}/avaliacoes": {
            "post": {
                "tags": [
                    "Avaliacoes"
                ],
                "summary": "Registra uma avaliação mútua após a conclusão da inscrição",
                "description": "ONG avalia o voluntário (nota 1–5); voluntário avalia a ONG. Cada lado pode avaliar uma única vez por inscrição.",
                "operationId": "24cf6e4e6fd4ba278ee9412d30015fe5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da inscrição concluída",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nota"
                                ],
                                "properties": {
                                    "nota": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comentario": {
                                        "type": "string",
                                        "example": "Excelente voluntário, pontual e dedicado."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Avaliação registrada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Inscrição não concluída ou sem permissão"
                    },
                    "404": {
                        "description": "Inscrição não encontrada"
                    },
                    "422": {
                        "description": "Avaliação já registrada ou dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/voluntarios/{id}/avaliacoes": {
            "get": {
                "tags": [
                    "Avaliacoes"
                ],
                "summary": "Lista as avaliações recebidas por um voluntário (feitas por ONGs)",
                "operationId": "3b44f5e07c4e5010c0b6730de652456c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de avaliações"
                    },
                    "404": {
                        "description": "Voluntário não encontrado"
                    }
                }
            }
        },
        "/api/ongs/{id}/avaliacoes": {
            "get": {
                "tags": [
                    "Avaliacoes"
                ],
                "summary": "Lista as avaliações recebidas por uma ONG (feitas por voluntários)",
                "operationId": "47db284356bd0e8ee8f7b6d5a5e17cc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de avaliações"
                    },
                    "404": {
                        "description": "ONG não encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticação via token Bearer (Laravel Sanctum)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
}